Add files via upload
here the computer will generate a random number from 0 to upper range which u will be asked,
and then u have to guess the number, if u get it right it will say u got it, if u don't it will give u certain hints,
and when u do get it , it will tell u no. of guesses u took.